Reorganize, remove unused logic
